Refactor: move nodes_mask_convertion nodes to nodes_mask.
